Make lfs_avail signed since it can temporarily go negative.
SCCS-vsn: sys/ufs/lfs/lfs.h 7.25

+++ usr/src/sys/ufs/lfs/lfs.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)lfs.h	7.24 (Berkeley) %G%
+ *	@(#)lfs.h	7.25 (Berkeley) %G%
  */
 
 #define	LFS_LABELPAD	8192		/* LFS label size */
@@ -70,7 +70,7 @@ struct lfs {
 	ino_t	lfs_free;		/* start of the free list */
 	u_long	lfs_bfree;		/* number of free disk blocks */
 	u_long	lfs_nfiles;		/* number of allocated inodes */
-	u_long	lfs_avail;		/* blocks available for writing */
+	long	lfs_avail;		/* blocks available for writing */
 	u_long  lfs_uinodes;		/* inodes in cache not yet on disk */
 	daddr_t	lfs_idaddr;		/* inode file disk address */
 	ino_t	lfs_ifile;		/* inode file inode number */
@@ -264,8 +264,8 @@ struct segsum {
  * the segment usage table, plus an ifile page.
  */
 #define LFS_FITS(fs, db)						\
-	((db + ((fs)->lfs_uinodes + INOPB((fs))) / INOPB((fs)) +	\
-	fsbtodb(fs, 1) + LFS_SUMMARY_SIZE / DEV_BSIZE + (fs)->lfs_segtabsz)\
+	((long)((db + ((fs)->lfs_uinodes + INOPB((fs))) / INOPB((fs)) +	\
+	fsbtodb(fs, 1) + LFS_SUMMARY_SIZE / DEV_BSIZE + (fs)->lfs_segtabsz))\
 	< (fs)->lfs_avail)
 
 /* Determine if a buffer belongs to the ifile */
